# CsharpApiSourceCode/GeneratedCSharpAPI_NEW_2/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Authorize.Net REST API",
    "description" : "Authorize.Net REST API",
    "contact" : {
      "name" : "developer_feedback@authorize.net"
    },
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "https://api.authorize.net/"
  } ],
  "paths" : {
...
    "/v1/webhooks" : {
      "get" : {
        "tags" : [ "webhooks-webhooks" ],
        "summary" : "List My Webhooks",
        "description" : "Retrieve details of all existing Webhooks for this account.",
        "operationId" : "getWebhooks",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/GetWebhookResponse"
                  },
                  "x-content-type" : "*/*"
                }
              }
            }
          },
          "500" : {
            "description" : "System Error.",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/errorResponse"
                }
              }
            }
          },
          "default" : {
            "description" : "Unexpected Error.",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/errorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "oauth" : [ "read" ]
        }, {
          "basic" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "webhooks-webhooks" ],
        "summary" : "Create A Webhook",
        "description" : "Enroll in one or more type of event notification.",
        "operationId" : "createWebhook",
        "requestBody" : {
          "description" : "The payload which contains the details for enrolling in one or more webhooks.",
          "content" : {
            "*/*" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateWebhookBody"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GetWebhookResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid/Missing Data or Limit Reached",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/errorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "System Error.",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/errorResponse"
                }
              }
            }
          },
          "default" : {
            "description" : "Unexpected Error.",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/errorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "oauth" : [ "read" ]
        }, {
          "basic" : [ ]
        } ],
        "x-codegen-request-body-name" : "createWebhookBody"
      }
    },
    "/v1/webhooks/{webhookId}" : {
      "get" : {
        "tags" : [ "webhooks-webhooks" ],
        "summary" : "Get a Webhook",
        "description" : "Retrieve THE details of an existing Webhook.",
        "operationId" : "getWebhook",
        "parameters" : [ {
          "name" : "webhookId",
          "in" : "path",
          "description" : "Unique identifier of the webhook. The identifier was sent in the response to the webhook request.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GetWebhookResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/errorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "System Error.",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/errorResponse"
                }
              }
            }
          },
          "default" : {
            "description" : "Unexpected Error.",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/errorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "oauth" : [ "read" ]
        }, {
          "basic" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "webhooks-webhooks" ],
        "summary" : "Update A Webhook",
        "description" : "Update the details of an existing webhook.",
        "operationId" : "updateWebhook",
        "parameters" : [ {
          "name" : "webhookId",
          "in" : "path",
          "description" : "Unique identifier of the webhook. The identifier was sent in the response to the webhook request.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "description" : "Contains the details to be updated.",
          "content" : {
            "*/*" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateWebhookBody"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GetWebhookResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid/Missing Data or Limit Reached",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/errorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/errorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "System Error.",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/errorResponse"
                }
              }
            }
          },
          "default" : {
            "description" : "Unexpected Error.",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/errorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "oauth" : [ "read" ]
        }, {
          "basic" : [ ]
        } ],
        "x-codegen-request-body-name" : "updateWebhookBody"
      },
      "delete" : {
        "tags" : [ "webhooks-webhooks" ],
        "summary" : "Delete A Webhook",
        "description" : "Delete an existing Webhook.",
        "operationId" : "deleteWebhook",
        "parameters" : [ {
          "name" : "webhookId",
          "in" : "path",
          "description" : "Unique identifier of the webhook. The identifier was sent in the response to the webhook request.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : { }
          },
          "400" : {
            "description" : "Invalid Data ",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/errorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/errorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "System Error.",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/errorResponse"
                }
              }
            }
          },
          "default" : {
            "description" : "Unexpected Error.",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/errorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "oauth" : [ "read" ]
        }, {
          "basic" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CreateWebhookBody" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "description" : "Contains details for the creation or update of a webhook."
          },
          "url" : {
            "type" : "string",
            "description" : "Callback URL for the webhook."
          },
          "eventTypes" : {
            "type" : "array",
            "description" : "List of events that will trigger notifications for this webhook. For a complete list of event types, see the [Webhooks API Documentation](https://developer.authorize.net/api/reference/features/webhooks.html#Event_Types_and_Payloads) page.",
            "items" : {
              "type" : "string"
            }
          },
          "status" : {
            "type" : "string",
            "description" : "New status of the Webhook."
          }
        },
        "description" : "Model for Create Webhook"
      },
      "GetWebhookResponse" : {
        "required" : [ "_links", "eventTypes", "status", "url", "webhookId" ],
        "type" : "object",
        "properties" : {
          "_links" : {
            "$ref" : "#/components/schemas/selfLink"
          },
          "webhookId" : {
            "type" : "string",
            "description" : "Unique identifier of the notification. The identifier was sent in the original notification."
          },
          "name" : {
            "type" : "string",
            "description" : "Label of the new Webhook."
          },
          "status" : {
            "type" : "string",
            "description" : "Unique identifier of the webhook. The identifier was sent in the response to the webhook request."
          },
          "url" : {
            "type" : "string",
            "description" : "Callback Url for the webhook."
          },
          "eventTypes" : {
            "type" : "array",
            "description" : "List of events that will trigger notifications for this webhook. For a complete list of event types, see the [Webhooks API Documentation](https://developer.authorize.net/api/reference/features/webhooks.html#Event_Types_and_Payloads) page.",
            "items" : {
              "type" : "string"
            }
          }
        },
        "description" : "Response model for get a Webhook",
        "example" : {
          "webhookId" : "webhookId",
          "_links" : {
            "self" : {
              "method" : "method",
              "href" : "href",
              "title" : "title"
            }
          },
          "name" : "name",
          "eventTypes" : [ "eventTypes", "eventTypes" ],
          "url" : "url",
          "status" : "status"
        }
      },
...
      "errorResponse" : {
        "title" : "errorResponse",
        "required" : [ "message", "reason" ],
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "string",
            "description" : "Describes the error condition."
          },
          "reason" : {
            "type" : "string",
            "description" : "This is an enum field that describes the business errors returned by the server. The client can use this field to interpret errors returned by the server. "
          },
          "informationLink" : {
            "type" : "string",
            "description" : "Contains a URL link to general online documentation for this error."
          },
          "correlationId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "details" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/errorDetails"
            }
          }
        },
        "description" : "Model for Error Response"
      },
      "errorDetails" : {
        "title" : "errorDetails",
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "string",
            "description" : "Describes the error condition."
          },
          "location" : {
            "type" : "string",
            "description" : "Indicates the name of a parameter."
          }
        },
        "description" : "Error details object."
      },
...
      "link" : {
        "type" : "object",
        "properties" : {
          "href" : {
            "type" : "string",
            "description" : "URI of the linked resource."
          },
          "title" : {
            "type" : "string",
            "description" : "Label of the linked resource."
          },
          "method" : {
            "type" : "string",
            "description" : "HTTP method of the linked resource."
          }
        },
        "example" : {
          "method" : "method",
          "href" : "href",
          "title" : "title"
        }
      },
...
      "selfLink" : {
        "type" : "object",
        "properties" : {
          "self" : {
            "$ref" : "#/components/schemas/link"
          }
        },
        "example" : {
          "self" : {
            "method" : "method",
            "href" : "href",
            "title" : "title"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "basic" : {
        "type" : "http",
        "scheme" : "basic"
      },
      "oauth" : {
        "type" : "oauth2",
        "flows" : {
          "implicit" : {
            "authorizationUrl" : "https://account.authorize.net/oauth/authorize",
            "scopes" : {
              "read" : "List and view details",
              "write" : "Create, list, and view details"
            }
          }
        }
      }
    }
  },
...
}